Split `serve:test_runtime_env`
