+++ python/ray/serve/BUILD
@@ -284,10 +284,10 @@ py_test(
 )
 
 py_test(
-    name = "test_runtime_env",
+    name = "test_runtime_env", "test_runtime_env_2"
     size = "large",
     srcs = serve_tests_srcs,
-    tags = ["exclusive", "post_wheel_build", "team:serve"],
+    tags = ["exclusive", "team:serve"],
     deps = [":serve_lib"],
 )
