+++ python/ray/serve/tests/conftest.py
@@ -1,10 +1,17 @@
 import os
 
 import pytest
+import tempfile
+import subprocess
+import random
 
 import ray
 from ray import serve
 from ray.serve.pipeline.generate import DeploymentNameGenerator
+
+# https://tools.ietf.org/html/rfc6335#section-6
+MIN_DYNAMIC_PORT = 49152
+MAX_DYNAMIC_PORT = 65535
 
 if os.environ.get("RAY_SERVE_INTENTIONALLY_CRASH", False) == 1:
     serve.controller._CRASH_AFTER_CHECKPOINT_PROBABILITY = 0.5
@@ -15,6 +22,35 @@ def ray_shutdown():
     yield
     serve.shutdown()
     ray.shutdown()
+
+
+@pytest.fixture
+def ray_start(scope="module"):
+    port = random.randint(MIN_DYNAMIC_PORT, MAX_DYNAMIC_PORT)
+    subprocess.check_output(
+        [
+            "ray",
+            "start",
+            "--head",
+            "--num-cpus",
+            "16",
+            "--ray-client-server-port",
+            f"{port}",
+        ]
+    )
+    try:
+        yield f"localhost:{port}"
+    finally:
+        subprocess.check_output(["ray", "stop", "--force"])
+
+
+@pytest.fixture
+def tmp_dir():
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        old_dir = os.getcwd()
+        os.chdir(tmp_dir)
+        yield tmp_dir
+        os.chdir(old_dir)
 
 
 @pytest.fixture(scope="session")
